requirements.txt
Django==4.2.4
django-environ==0.10.0
django-extensions==3.2.3
Werkzeug==2.3.6
gunicorn==21.2.0
Pillow==10.0.0
psycopg[binary]==3.1.10
awesome-slugify==1.6.5
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
# move the below to dev requirements
factory-boy==3.3.0
pytest==7.4.0
pytest-django==4.5.2
pytest-socket==0.6.0
pytest-timeout==2.1.0
pytest-xdist==3.3.1
ipdb==0.13.13
